Add PERL method to evolve the system
Adapt the ODE of the system, now it calculates the acceleration of each one of three bodies explicitly.
Add a function to calculate the energy and angular momentum of the system. It's optional.
Change from Verlet method to PEFRL method to the system's evolution.
Adapt function to create images, now the masses leave a line over their journey.
Solve the problem in the creation of video, when the libraries were importing.